Function quinn_ffi::ffi::last_error

#[no_mangle]
pub unsafe extern "cdecl" fn last_error(
    error_buf: Out<'_, u8>, 
    error_buf_len: size_t, 
    actual_error_buf_len: Out<'_, size_t>
) -> FFIResult

Retrieves the last occurred error.

error_buf: Allocated memory for the error message destination.

error_buf_len: The size of the allocated error message buffer error_buf.

actual_error_buf_len: Allocated memory for the actual length of the error buffer.

actual_error_buf_len could be used to resize buffer if result returns BufferToSmall.