Module quinn_ffi::proto_impl

Quinn-proto implementation, similar to quinn but without the async runtime and some differences to make it fit with FFi applications.

Structs

ConnectionImpl

A QUIC connection using quinn-proto.

EndpointImpl

A QUIC endpoint using quinn-proto.

EndpointPoller

Polls the endpoint when notified to do so. This polling happens on its own thread.

IpAddr

IpAddress that is FFI safe.

Enums

ConnectionEvent

Events for the connection.

EndpointEvent

Events for the endpoint.

FFIErrorKind

An Error implementing type that can be returned in a Result.